// src/types/tmdb.types.ts
export type TMovie = {
  id: number;
  adult: boolean;
  backdrop_path: string;
  genre_ids: number[];
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: string;
  release_date: string; // format YYYY-MM-DD
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
};